+++ ML/Mathematics/MathUtils.cs
@@ -16,7 +16,7 @@ namespace ML.Mathematics
     /// Calculates h(z) = -z log2(z)
     /// </summary>
     [MethodImpl(MethodImplOptions.AggressiveInlining)]
-    public static double EntropyH(float z)
+    public static double EntropyH(double z)
     {
       const double COEFF = 1.44269504089F; // 1/ln(2)
       return (0.0D <= z && z < double.Epsilon) ? 0.0D : -z*Math.Log(z)*COEFF;
